fix(mixin): server level tick block mixin not def class net/minecraft/server/level/ServerLevel$Anonymous$f037024943ef4d1a8b86e17b3519f291.
